# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Leitstand API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/missions/{mission_id}/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Mission Runs",
        "description": "List a mission's runs, newest first, without their plans.\n\nEvery execution the mission has ever had, whatever its outcome. Each carries its own robot,\ntimings, status and a digest of the plan it executed, so runs of an edited mission can be\ntold apart from runs of the plan as it stands.",
        "operationId": "list_mission_runs",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only runs created before this instant; pages backwards.",
              "title": "Before"
            },
            "description": "Only runs created before this instant; pages backwards."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryView"
                  },
                  "title": "Response List Mission Runs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "description": "Return one run with the plan it executed, frozen at dispatch, and its failure cause.",
        "operationId": "get_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "runs"
        ],
        "summary": "Annotate Run",
        "description": "Set or clear a run's notes. Notes are the operator's, editable at any time.",
        "operationId": "annotate_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunNotesPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "runs"
        ],
        "summary": "Delete Run",
        "description": "Remove a finished run and its per-stage record.\n\nA run still active is refused (409): its robot is driving it. The audit log keeps every\ncompleted delete; what goes is the detail.",
        "operationId": "delete_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/state": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run State",
        "description": "Return one run's per-stage state with errors attributed per stage.\n\nA run the robot has not reported on yet projects every stage as WAITING from its own plan.",
        "operationId": "get_run_state",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStateView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoverageStage": {
        "properties": {
          "stage_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stage Id"
          },
          "on_cancel": {
            "anyOf": [
              {
                "items": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/NavigationStage"
                    },
                    {
                      "$ref": "#/components/schemas/CoverageStage"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "coverage": "#/components/schemas/CoverageStage",
                      "navigation": "#/components/schemas/NavigationStage"
                    }
                  }
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Cancel",
            "description": "Cleanup stages executed sequentially when this stage is cancelled. Cleanup stages are themselves non-cancellable."
          },
          "kind": {
            "type": "string",
            "const": "coverage",
            "title": "Kind",
            "default": "coverage"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/Segment"
            },
            "type": "array",
            "minItems": 1,
            "title": "Segments",
            "description": "The whole stage as one ordered route. Concatenating the segments gives the drivable line, turns included; the kinds say which of it is a worked swath. Consecutive segments share an endpoint, which a receiver joining them skips."
          },
          "provenance": {
            "$ref": "#/components/schemas/CoverageProvenance",
            "description": "The inputs, measurements and act that produced these segments."
          }
        },
        "type": "object",
        "required": [
          "stage_id",
          "segments",
          "provenance"
        ],
        "title": "CoverageStage",
        "description": "Cover a field by driving its swaths in order, each reached by the turn before it.\n\nProvenance is required, so this stage cannot be written by hand: swaths come from the\nplanner, and the checks that refuse a robot which would cut the corners read the parameters\nthe plan was made with."
      },
...
      "ErrorOrigin": {
        "type": "string",
        "enum": [
          "robot",
          "backend"
        ],
        "title": "ErrorOrigin",
        "description": "Source of a mission error: the robot's execution, or a backend-authored cause."
      },
...
      "ErrorSeverity": {
        "type": "string",
        "enum": [
          "WARNING",
          "FATAL"
        ],
        "title": "ErrorSeverity",
        "description": "Severity of an error reported by the robot."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionError": {
        "properties": {
          "origin": {
            "$ref": "#/components/schemas/ErrorOrigin"
          },
          "severity": {
            "$ref": "#/components/schemas/ErrorSeverity"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Stable identifier for programmatic dispatch (e.g. ``boundary_anchor_mismatch``)."
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ErrorReference"
            },
            "type": "array",
            "title": "References"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "origin",
          "severity",
          "type",
          "description"
        ],
        "title": "MissionError",
        "description": "A structured error report attached to a stage or mission.\n\nMirrors the proto ``Error`` (severity / type / references / description); ``origin``\nis a backend-only marker distinguishing a robot fault from a backend-authored cause."
      },
...
      "NavigationStage": {
        "properties": {
          "stage_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stage Id"
          },
          "on_cancel": {
            "anyOf": [
              {
                "items": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/NavigationStage"
                    },
                    {
                      "$ref": "#/components/schemas/CoverageStage"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "coverage": "#/components/schemas/CoverageStage",
                      "navigation": "#/components/schemas/NavigationStage"
                    }
                  }
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Cancel",
            "description": "Cleanup stages executed sequentially when this stage is cancelled. Cleanup stages are themselves non-cancellable."
          },
          "kind": {
            "type": "string",
            "const": "navigation",
            "title": "Kind",
            "default": "navigation"
          },
          "waypoints": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/WGS84Waypoint"
                },
                {
                  "$ref": "#/components/schemas/SiteLocalWaypoint"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "site_local": "#/components/schemas/SiteLocalWaypoint",
                  "wgs84": "#/components/schemas/WGS84Waypoint"
                }
              }
            },
            "type": "array",
            "minItems": 1,
            "title": "Waypoints",
            "description": "Ordered waypoints to traverse. All waypoints in one stage must share their ``kind`` (homogeneity); this is enforced by the backend at dispatch, not by this schema."
          }
        },
        "type": "object",
        "required": [
          "stage_id",
          "waypoints"
        ],
        "title": "NavigationStage",
        "description": "Drive the robot through an ordered list of waypoints."
      },
...
      "RunNotesPatch": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "RunNotesPatch"
      },
      "RunOriginView": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "manual",
              "agent"
            ],
            "title": "Kind"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Id"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "actor"
        ],
        "title": "RunOriginView"
      },
...
      "RunSiteAnchor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "anchor_lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Anchor Lat"
          },
          "anchor_lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Anchor Lon"
          },
          "anchor_heading_deg": {
            "type": "number",
            "exclusiveMaximum": 360.0,
            "minimum": 0.0,
            "title": "Anchor Heading Deg"
          }
        },
        "type": "object",
        "required": [
          "name",
          "anchor_lat",
          "anchor_lon",
          "anchor_heading_deg"
        ],
        "title": "RunSiteAnchor",
        "description": "Where a site stood when a run was dispatched into it.\n\nA site-local waypoint is a pair of numbers that mean nothing without the frame they are\nmeasured in, so the frame is copied into the run beside the plan. The name is copied\ntoo, so the run stays readable after the site is deleted."
      },
      "RunStateView": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "mission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mission Id"
          },
          "stage_states": {
            "items": {
              "$ref": "#/components/schemas/StageStateView"
            },
            "type": "array",
            "title": "Stage States"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "mission_id",
          "stage_states"
        ],
        "title": "RunStateView",
        "description": "A run's per-stage state as one read shape for both REST and the live WS frame.\n\nCarries both ids: the WS topic is keyed by mission, so a subscriber showing one run must\nfilter frames by ``run_id``, and a list of missions indexes them by ``mission_id``."
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "DISPATCHED",
          "RUNNING",
          "PAUSED",
          "SUCCEEDED",
          "FAILED",
          "CANCELLED",
          "REJECTED"
        ],
        "title": "RunStatus",
        "description": "Lifecycle state of a MissionRun, owned by the backend.\n\n``PENDING`` is the run row committed before the robot has answered the dispatch. ``REJECTED``\nis kept apart from ``FAILED`` because they mean different things for the robot: a rejection\nmeans it is definitely not driving, while a timeout or a mid-run failure means it may be."
      },
      "RunSummaryView": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "mission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mission Id"
          },
          "robot_id": {
            "type": "string",
            "title": "Robot Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stages_digest": {
            "type": "string",
            "title": "Stages Digest",
            "description": "Content digest of what this run was told to do, ignoring stage ids and how a stage was produced. Two runs with the same digest drove the same ground."
          },
          "origin": {
            "$ref": "#/components/schemas/RunOriginView"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "dispatched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatched At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "last_frame_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Frame At",
            "description": "When the robot last reported on this run, by the backend's clock."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "When the backend last changed this run: a status transition or a notes edit."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "mission_id",
          "robot_id",
          "status",
          "stages_digest",
          "origin",
          "notes",
          "created_at",
          "dispatched_at",
          "started_at",
          "ended_at",
          "last_frame_at",
          "updated_at"
        ],
        "title": "RunSummaryView",
        "description": "A run without its stages: what a list shows, and what a mission carries as ``latest_run``."
      },
      "RunView": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "mission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mission Id"
          },
          "robot_id": {
            "type": "string",
            "title": "Robot Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "stages_digest": {
            "type": "string",
            "title": "Stages Digest",
            "description": "Content digest of what this run was told to do, ignoring stage ids and how a stage was produced. Two runs with the same digest drove the same ground."
          },
          "origin": {
            "$ref": "#/components/schemas/RunOriginView"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "dispatched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatched At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "last_frame_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Frame At",
            "description": "When the robot last reported on this run, by the backend's clock."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "When the backend last changed this run: a status transition or a notes edit."
          },
          "stages": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/NavigationStage"
                },
                {
                  "$ref": "#/components/schemas/CoverageStage"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "coverage": "#/components/schemas/CoverageStage",
                  "navigation": "#/components/schemas/NavigationStage"
                }
              }
            },
            "type": "array",
            "title": "Stages"
          },
          "site_anchors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/RunSiteAnchor"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Anchors",
            "description": "The frame of each site this run drove in, as it stood at dispatch, keyed by site id. Site-local waypoints are measured against these rather than against the site as it is now, so moving or deleting a site cannot move a run that already happened."
          },
          "failure_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MissionError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Errors"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "mission_id",
          "robot_id",
          "status",
          "stages_digest",
          "origin",
          "notes",
          "created_at",
          "dispatched_at",
          "started_at",
          "ended_at",
          "last_frame_at",
          "updated_at",
          "stages"
        ],
        "title": "RunView",
        "description": "A run with the stages it was dispatched with, frozen as they stood, and any failure."
      },
...
      "StageStateView": {
        "properties": {
          "stage_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stage Id"
          },
          "stage_index": {
            "type": "integer",
            "title": "Stage Index"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/MissionError"
            },
            "type": "array",
            "title": "Errors"
          },
          "status_source": {
            "type": "string",
            "enum": [
              "robot",
              "backend"
            ],
            "title": "Status Source",
            "description": "Who established the status: 'robot' when the robot reported it, 'backend' when the backend projected it because the run ended before the robot reported this stage."
          }
        },
        "type": "object",
        "required": [
          "stage_id",
          "stage_index",
          "status",
          "progress",
          "started_at",
          "ended_at",
          "result",
          "errors",
          "status_source"
        ],
        "title": "StageStateView",
        "description": "One stage's runtime state plus the errors attributed to it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WGS84Waypoint": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "wgs84",
            "title": "Kind",
            "default": "wgs84"
          },
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "description": "Degrees latitude, WGS84."
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "description": "Degrees longitude, WGS84."
          },
          "heading_deg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMaximum": 360.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading Deg",
            "description": "Target robot heading in degrees clockwise from true north (compass bearing), [0, 360)."
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "WGS84Waypoint",
        "description": "A waypoint anchored in the WGS84 geographic frame."
      },
...
    }
  }
}